Separated discover and Top Plays
