+++ src/components/TopArtists.jsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React from "react";
 import Loader from "./Loader";
 import ErrorComponent from "./ErrorComponent";
 import ArtistCard from "./ArtistCard";
-import { useGetTopSongsQuery } from '../redux/APIs/ShazamApi1';
+import { useGetTopSongsQuery } from "../redux/APIs/ShazamApi1";
 
 const TopArtists = () => {
   const { data: topSongs, isFetching, isError } = useGetTopSongsQuery();
@@ -12,11 +12,17 @@ const TopArtists = () => {
   if (isError) return <ErrorComponent />;
 
   return (
-    <div className="flex flex-col m-4">
-      <h2 className="font-bold text-3xl text-white text-left mt-4 mb-10">Top artists</h2>
+    <div className="flex flex-col m-4 w-fit">
+      <div className="flex flex-col m-4">
+        <h2 className="font-bold text-3xl text-white text-left mt-4 mb-10">
+          Top Artists
+        </h2>
 
-      <div className="flex flex-wrap justify-center gap-8">
-        {topSongs?.tracks?.map((track) => <ArtistCard key={track?.key} track={track} />)}
+        <div className="flex flex-wrap justify-center gap-8">
+          {topSongs?.tracks?.map((track) => (
+            <ArtistCard key={track?.key} track={track} />
+          ))}
+        </div>
       </div>
     </div>
   );
